+++ src/ring_buffer/variants/alloc_ext.rs
@@ -1,24 +1,35 @@
 #![cfg(feature = "alloc")]
 use alloc::vec;
 use alloc::vec::Vec;
-use crate::{ConcurrentHeapRB, LocalHeapRB};
+use crate::{ConcurrentHeapRB, LocalHeapRB, UnsafeSyncCell};
 use crate::HeapStorage;
+#[allow(unused_imports)]
+use crate::ProdIter;
 
 // Concurrent
 
 impl<T> From<Vec<T>> for ConcurrentHeapRB<T> {
-    fn from(value: Vec<T>) -> Self {
+    /// Constructs a `ConcurrentHeapRB` using the passed vector.
+    /// Note that the length of the buffer will be equal to the length of teh vector, and *not*
+    /// to its capacity.
+    fn from(mut value: Vec<T>) -> Self {
+        value.shrink_to_fit();
         Self::_from(HeapStorage::from(value))
     }
 }
 
 impl<T> ConcurrentHeapRB<T> {
-    /// Creates a new concurrent heap-allocated buffer with given capacity and uninitialised elements.
-    #[allow(clippy::uninit_vec)] // handled later by UnsafeSyncCell
-    pub fn new(capacity: usize) -> Self {
+    /// Creates a new concurrent heap-allocated buffer with given capacity and zeroed (uninitialised) elements.
+    /// # Safety
+    /// The buffer must be then initialised using proper [`ProdIter`] methods (`*_init` ones).
+    pub unsafe fn new_zeroed(capacity: usize) -> Self {
         let mut v = Vec::with_capacity(capacity);
-        unsafe { v.set_len(capacity); }
-        Self::from(v)
+
+        for _ in 0..capacity {
+            v.push(UnsafeSyncCell::new_zeroed());
+        }
+
+        Self::_from(HeapStorage::from(v))
     }
 
     /// Creates a new concurrent heap-allocated buffer with given capacity and elements initialised to `default`.
@@ -32,18 +43,27 @@ impl<T> ConcurrentHeapRB<T> {
 // Local
 
 impl<T> From<Vec<T>> for LocalHeapRB<T> {
-    fn from(value: Vec<T>) -> Self {
+    /// Constructs a `LocalHeapRB` using the passed vector.
+    /// Note that the length of the buffer will be equal to the length of teh vector, and *not*
+    /// to its capacity.
+    fn from(mut value: Vec<T>) -> Self {
+        value.shrink_to_fit();
         Self::_from(HeapStorage::from(value))
     }
 }
 
 impl<T> LocalHeapRB<T> {
-    /// Creates a new local heap-allocated buffer with given capacity.
-    #[allow(clippy::uninit_vec)] // handled later by UnsafeSyncCell
-    pub fn new(capacity: usize) -> Self {
+    /// Creates a new local heap-allocated buffer with given capacity and zeroed (uninitialised) elements.
+    /// # Safety
+    /// The buffer must be then initialised using proper [`ProdIter`] methods (`*_init` ones).
+    pub unsafe fn new_zeroed(capacity: usize) -> Self {
         let mut v = Vec::with_capacity(capacity);
-        unsafe { v.set_len(capacity); }
-        Self::from(v)
+
+        for _ in 0..capacity {
+            v.push(UnsafeSyncCell::new_zeroed());
+        }
+
+        Self::_from(HeapStorage::from(v))
     }
 
     /// Creates a new local heap-allocated buffer with given capacity and elements initialised to `default`.