+++ tests/datached_work_tests.rs
@@ -40,7 +40,7 @@ fn prepare(mut prod: ProdIter<ConcurrentHeapRB<usize>>, mut work: WorkIter<Concu
 
 #[test]
 fn test_work_detached_sync_index() {
-    let (prod, work, cons) = ConcurrentHeapRB::new(BUFFER_SIZE + 1).split_mut();
+    let (prod, work, cons) = ConcurrentHeapRB::default(BUFFER_SIZE + 1).split_mut();
 
     let (mut prod, mut work, mut cons) = prepare(prod, work, cons);
 
@@ -51,7 +51,7 @@ fn test_work_detached_sync_index() {
     assert_eq!(cons.available(), BUFFER_SIZE);
 
     for i in 0..BUFFER_SIZE {
-        assert_eq!(cons.pop().unwrap(), i + 1);
+        unsafe { assert_eq!(cons.pop().unwrap(), i + 1); }
     }
 
     assert_eq!(prod.available(), BUFFER_SIZE);
@@ -61,7 +61,7 @@ fn test_work_detached_sync_index() {
 
 #[test]
 fn test_work_detached() {
-    let (prod, work, cons) = ConcurrentHeapRB::new(BUFFER_SIZE + 1).split_mut();
+    let (prod, work, cons) = ConcurrentHeapRB::default(BUFFER_SIZE + 1).split_mut();
 
     let (mut prod, work, mut cons) = prepare(prod, work, cons);
 
@@ -72,7 +72,7 @@ fn test_work_detached() {
     assert_eq!(cons.available(), BUFFER_SIZE);
 
     for i in 0..BUFFER_SIZE {
-        assert_eq!(cons.pop().unwrap(), i + 1);
+        unsafe { assert_eq!(cons.pop().unwrap(), i + 1); }
     }
 
     assert_eq!(prod.available(), BUFFER_SIZE);
@@ -82,7 +82,7 @@ fn test_work_detached() {
 
 #[test]
 fn test_work_detached_set_index() {
-    let (prod, work, cons) = ConcurrentHeapRB::new(BUFFER_SIZE + 1).split_mut();
+    let (prod, work, cons) = ConcurrentHeapRB::default(BUFFER_SIZE + 1).split_mut();
 
     let (mut prod, mut work, mut cons) = prepare(prod, work, cons);
 
@@ -99,7 +99,7 @@ fn test_work_detached_set_index() {
     assert_eq!(cons.available(), BUFFER_SIZE - 1);
 
     for i in 0..BUFFER_SIZE - 1 {
-        assert_eq!(cons.pop().unwrap(), i + 1);
+        unsafe { assert_eq!(cons.pop().unwrap(), i + 1); }
     }
 
     assert_eq!(prod.available(), BUFFER_SIZE - 1);
@@ -109,7 +109,7 @@ fn test_work_detached_set_index() {
 
 #[test]
 fn test_work_go_back() {
-    let buf: ConcurrentHeapRB<usize> = ConcurrentHeapRB::new(BUFFER_SIZE + 1);
+    let buf: ConcurrentHeapRB<usize> = ConcurrentHeapRB::default(BUFFER_SIZE + 1);
     let (_, work, _) = buf.split_mut();
 
     let mut work = work.detach();